requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
regex==2024.5.15
dateparser==1.2.0
PyYAML==6.0.2
requests==2.32.3
